requirement.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
